requirements.txt
torch>=2.0.1
torchvision>=0.15.2
Pillow>=10.0.1
matplotlib>=3.8.0
numpy>=1.23.5